Refactor CLI code into separate input, output and settings modules
