+++ openhands/core/cli_settings.py
@@ -1,10 +1,12 @@
 from prompt_toolkit import PromptSession, print_formatted_text
 from prompt_toolkit.completion import FuzzyWordCompleter
 from prompt_toolkit.formatted_text import HTML
+from prompt_toolkit.shortcuts import print_container
+from prompt_toolkit.widgets import Frame, TextArea
 from pydantic import SecretStr
 
-import openhands.agenthub  # noqa F401 (we import this to get the agents registered)
 from openhands.controller.agent import Agent
+from openhands.core.cli_output import COLOR_GREY
 from openhands.core.cli_prompts import UserCancelledError, cli_confirm, kb_cancel
 from openhands.core.cli_utils import (
     VERIFIED_ANTHROPIC_MODELS,
@@ -21,6 +23,80 @@ from openhands.memory.condenser.impl.llm_summarizing_condenser import (
 from openhands.storage.data_models.settings import Settings
 from openhands.storage.settings.file_settings_store import FileSettingsStore
 from openhands.utils.llm import get_supported_llm_models
+
+
+def display_settings(config: AppConfig):
+    llm_config = config.get_llm_config()
+    advanced_llm_settings = True if llm_config.base_url else False
+
+    # Prepare labels and values based on settings
+    labels_and_values = []
+    if not advanced_llm_settings:
+        # Attempt to determine provider, fallback if not directly available
+        provider = getattr(
+            llm_config,
+            'provider',
+            llm_config.model.split('/')[0] if '/' in llm_config.model else 'Unknown',
+        )
+        labels_and_values.extend(
+            [
+                ('   LLM Provider', str(provider)),
+                ('   LLM Model', str(llm_config.model)),
+                ('   API Key', '********' if llm_config.api_key else 'Not Set'),
+            ]
+        )
+    else:
+        labels_and_values.extend(
+            [
+                ('   Custom Model', str(llm_config.model)),
+                ('   Base URL', str(llm_config.base_url)),
+                ('   API Key', '********' if llm_config.api_key else 'Not Set'),
+            ]
+        )
+
+    # Common settings
+    labels_and_values.extend(
+        [
+            ('   Agent', str(config.default_agent)),
+            (
+                '   Confirmation Mode',
+                'Enabled' if config.security.confirmation_mode else 'Disabled',
+            ),
+            (
+                '   Memory Condensation',
+                'Enabled' if config.enable_default_condenser else 'Disabled',
+            ),
+        ]
+    )
+
+    # Calculate max widths for alignment
+    # Ensure values are strings for len() calculation
+    str_labels_and_values = [(label, str(value)) for label, value in labels_and_values]
+    max_label_width = (
+        max(len(label) for label, _ in str_labels_and_values)
+        if str_labels_and_values
+        else 0
+    )
+
+    # Construct the summary text with aligned columns
+    settings_lines = [
+        f'{label+":":<{max_label_width+1}} {value:<}'  # Changed value alignment to left (<)
+        for label, value in str_labels_and_values
+    ]
+    settings_text = '\n'.join(settings_lines)
+
+    container = Frame(
+        TextArea(
+            text=settings_text,
+            read_only=True,
+            style=COLOR_GREY,
+            wrap_lines=True,
+        ),
+        title='Settings',
+        style=f'fg:{COLOR_GREY}',
+    )
+
+    print_container(container)
 
 
 async def modify_llm_settings_basic(
